aadfsdf

+++ ft_strlen.c
@@ -6,11 +6,13 @@
 /*   By: vipereir <vipereir@student.42.rio>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/29 18:36:45 by vipereir          #+#    #+#             */
-/*   Updated: 2022/05/04 11:34:29 by vipereir         ###   ########.fr       */
+/*   Updated: 2022/05/10 11:54:17 by vipereir         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int	ft_strlen(const char *s)
+#include <stdlib.h>
+
+size_t	ft_strlen(const char *s)
 {
 	int		i;
